feat: Implement BaseConfigManager for credential management
- Added BaseConfigManager class to handle configuration directory and file operations.
- Implemented methods for reading, writing, and managing user credentials.
- Introduced error handling for file operations and credential retrieval.

feat: Create ConfigDisplayManager for displaying credentials

- Added ConfigDisplayManager class to format and display user credentials.
- Implemented methods to mask sensitive information and present configuration summaries.
- Enhanced output with color-coded formatting using chalk.

feat: Develop InteractiveConfigManager for user input

- Introduced InteractiveConfigManager class to facilitate interactive user prompts.
- Implemented methods for gathering user input on profiles, providers, models, and settings.
- Added specific methods for AWS and Azure configurations.

feat: Add ProviderConfigManager for provider-specific configurations

- Created ProviderConfigManager class to manage provider-specific settings.
- Implemented methods for configuring base URLs for various AI providers.
- Enhanced user guidance with notes on default values and requirements.
